pyproject.toml
[project]
name = "monitor-fast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.124.2",
    "fastapi-utils>=0.8.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "scalar-fastapi>=1.5.0",
    "tortoise-orm>=0.25.1",
    "typing-inspect>=0.9.0",
    "uvicorn>=0.38.0",
]
